requirements.txt
streamlit
ultralytics
opencv-python-headless
easyocr
pillow
numpy
